[PORT FROM MAIN]wifi: dns forwarders are not set after dnsmasq is restarted
BZ: 29453 25452 17650

Patch set 2: a counter is added to usleep and the duration of the sleep is
decreased to 1ms. We can use a synchronization mechanism instead of the
sleep (using shared semaphore between dnsmasq and netd processes, or
by modifying the existing pipe between netd and dnsmasq to report events
from dnsmasq to netd... but this is more complicated than simple sleep).

The actual netd tethering implementation restarts the dnsmasq when
a new tethered interface goes up or down, but the dns servers are not
set (using update_dns command).
Fix: update_dns using the mDnsForwarders servers list is called
after the dnsmasq is restarted in (un)tetherInterface function.
Note: 1- usleep(10000) is used to wait for dnsmasq to be started.
      2- We could call "tether dns set" (NetworkManagement::
SetDnsForwarders) from the Tethering java code, but keeping the
dns update at the netd level is simpler and keeps the dnsmasq restart
hidden to the upper levels.

Fix klocwork issue

Change-Id: I88fe8934a31d368a907598abad9d2b0c1c614648
Orig-Change-Id: I7b64777f4baf416b3a5e270d2c91f8ede0163115
Category: device enablement
Domain: CWS.WIFI-Common
Origin: upstream
Upstream-Candidate: yes
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>