TetherController: dns forwarders are not set after dnsmasq is restarted.
BZ: 29453 25452 17650

Patch set 2: a counter is added to usleep and the duration of the sleep is
decreased to 1ms. We can use a synchronization mechanism instead of the
sleep (using shared semaphore between dnsmasq and netd processes, or
by modifying the existing pipe between netd and dnsmasq to report events
from dnsmasq to netd... but this is more complicated than simple sleep).

The actual netd tethering implementation restarts the dnsmasq when
a new tethered interface goes up or down, but the dns servers are not
set (using update_dns command).
Fix: update_dns using the mDnsForwarders servers list is called
after the dnsmasq is restarted in (un)tetherInterface function.
Note: 1- usleep(10000) is used to wait for dnsmasq to be started.
      2- We could call "tether dns set" (NetworkManagement::
SetDnsForwarders) from the Tethering java code, but keeping the
dns update at the netd level is simpler and keeps the dnsmasq restart
hidden to the upper levels.

Fix klocwork issue

Change-Id: I482b2433aec626e7caf644e4f6886131c888e6b1
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Reviewed-on: http://android.intel.com:8080/78499
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>