[PORT FROM MAIN]wifi: Reset the InterfaceTable upon failure
BZ: 24583

Since the MOBILE_HIPRI is not a default connection,
ConnectivityService adds the route information of the
established HIPRI connection as a secondary one. Adding
of the route to the SecondaryTable failed. Eventhough the
adding of route information fails, interface name maintained in
the mInterfaceTable structure is not reset. When the call to
enableNat is made by Tethering service, NatController first
checks whether it's a secondary connection. Since the interface
table in the SecondaryTableController is not reset, NatController
tries to do the operations on the invalid secondary interface
resulting in exception thrown and eventually resulting in stopping
the tethering service.

This patch resets the mInterfaceTable upon failure.

Change-Id: I026068405ddf33d0bbcdcdeb693bf86d4ea646dd
Orig-Change-Id: I467238eb1da7d057ba7cc5f16c4a51bf6b22a04d
Category: device enablement
Domain: CWS.WIFI-Common
Origin: upstream
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>