[PORT FROM MAIN] tethering/widi: Limit DHCP server network interface coverage
BZ: 17348

The use case USB tethering/3G + WiDi was not working because of conflicts
between  DHCP servers.

The problem is that the Tethering DHCP server covers all the network
interfaces instead of covering only the Tethered interfaces.

This patches modifies netd in order the tethering and WiDi DHCP servers
to cover only the necessary interfaces, and thus beeing able to activate
the scenario USB tethering/3G + widi.

Change-Id: Ic73e6b9a815566a16c1dd3a079ef0ac69f4c4c84
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: http://android.intel.com:8080/77539
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>