WiFi: Reset netd LOCAL_MODULE to origin name.
BZ: 97757

By default, netd is built. It is listed in
build/target/product/base.mk.
So rename netd.bcm to netd in order to compile it
by default. It is required to have it in the board
to be able to boot. Even if no WiFi module is selected.
In TI case, the compile flag "wl12xx-compat" select the
correct netd.ti instead of netd.

Change-Id: I0078834804d4979462b3357a74c2b859cd62dd50
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: http://android.intel.com:8080/100743
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>