[PORT FROM MAIN] wifi: Reset netd LOCAL_MODULE to origin name.
BZ: 97757

By default, netd is built. It is listed in
build/target/product/base.mk.
So rename netd.bcm to netd in order to compile it
by default. It is required to have it in the board
to be able to boot. Even if no WiFi module is selected.
In TI case, the compile flag "wl12xx-compat" select the
correct netd.ti instead of netd.

Feature   : Wlan
Category  : Improvement
Purpose   : Scalability
Origin    : Upstream
Upstream-candidate: no

Change-Id: I060b58f42189824bfb94d94c39d572148493a52e
Orig-Change-Id: I0078834804d4979462b3357a74c2b859cd62dd50
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>