WiFi Hotspot: Enable 5GHz and offer advanced settings to the user
BZ: 120839 123820

This patch allows the user to configure the WiFi Hotspot in an advanced way. The
user can now choose the band (2.4 or 5 GHz), change the hw mode ('bg', 'a' or
'ac'), set the 80211n mode and select the radio channel according to local
regulatory.
The UI is also ready to act on routing parameters like IP address and subnet
mask.

Technically this patch:
- Set the hostapd supplicant with hw, 80211n, channel and country parameters.
- Save and retrieve user choices with ConfigStore.
- Get from driver the allowed radio channels according to local regulatory.
- Reset the radio parameters when they are no more applicable in the new locale.

Change-Id: I02cb77b7c6169f9a1016954f939dbebb857cb206
Signed-off-by: Jonathan DE CESCO <jonathan.de.cesco@intel.com>
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>